refactor(init): remove internal names from root __init__.py
This removes internal names from the root __init__.py and refactors the rest of the code so that
the full paths to these names are used, rather than relying on them being in the root __init__.py.
