+++ docs/api/source/trame.rst
@@ -8,5 +8,5 @@ trame
 .. automodule:: trame
   :members:
   :undoc-members:
-  :exclude-members: activate_app, base_directory, create_app, deactivate_app, enable_module, get_app_instance, get_version, js_call, js_property, log_js_error, main, print_server_info, trigger_key, validate_key_names
+  :exclude-members:
   :show-inheritance:
