+++ setup.cfg
@@ -31,4 +31,4 @@ include_package_data = True
 
 [options.entry_points]
 console_scripts =
-    trame = trame:main
+    trame = trame.dev:main
