+++ trame/__init__.py
@@ -1,15 +1,10 @@
-from trame.app import get_app_instance
-from trame.dev import main
 from trame.layouts import update_layout
 from trame.server import port, start, stop
 from trame.state import (
     change, flush_state, get_state, is_dirty, is_dirty_all, State, update_state
 )
-from trame.trigger import Controller, trigger, trigger_key
-from trame.utils import (
-    get_cli_parser, get_version, log_js_error, print_server_info,
-    validate_key_names
-)
+from trame.trigger import Controller, trigger
+from trame.utils import get_cli_parser, get_version
 
 __version__ = get_version()
 
@@ -79,11 +74,4 @@ __all__ = [
 
     # These are not exposed in the docs
     "__version__",
-    "get_app_instance",
-    "get_version",
-    "log_js_error",
-    "main",
-    "print_server_info",
-    "trigger_key",
-    "validate_key_names",
 ]
