+++ trame/html/vuetify/__init__.py
@@ -1,4 +1,4 @@
-from trame import get_app_instance
+from trame.app import get_app_instance
 from trame.html import AbstractElement, Template
 
 try:
